Refactor dir hierarchy
