+++ e2e/node/server/express.ts
@@ -22,10 +22,10 @@ import log from "loglevel";
 import express from "express";
 // Here we want to test how the local code behaves, not the already published one.
 // eslint-disable-next-line import/no-relative-packages
-import { Session } from "../../packages/node/src/index";
+import { Session } from "../../../packages/node/src/index";
 // Extensions are required for JSON-LD imports.
 // eslint-disable-next-line import/extensions
-import CONSTANTS from "../../playwright.client-authn.constants.json";
+import CONSTANTS from "../../../playwright.client-authn.constants.json";
 
 log.setLevel("TRACE");
